requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
channels>=4.0
channels-redis>=4.2
daphne>=4.1
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
django-environ>=0.11
django-cors-headers>=4.3
argon2-cffi>=23.1
gunicorn>=21.2
uvicorn>=0.27
homeassistant-api
paho-mqtt>=1.6
cryptography>=42.0
websocket-client>=1.8
